refactor: move component definition into BodyTemplate.tsx
